feat: Use immer for mutations
It can be tricky to perform mutations on complex state because it
requires to perform shallow copy on many levels, as a shallow copy keep
references.
Thus, we use Immer, which is also used behind the scene by Redux.
It greatly eases the state mutations, allowing much simpler code.
Too bad we didn't use it before...