feat: Implement undo action on back button
When the user clicks on the back button, the last action is undone.
This is done through undo-redux that handle the logic.
Note a small refacto was necessary, so the last card revealing is no
longer a dedicated action but rather handled in the moveCard reducer
method, as this was troublesome when undoing action.